Move ConfigRevision resources from extras to core
