+++ netbox/core/tables/__init__.py
@@ -1,2 +1,3 @@
+from .config import *
 from .data import *
 from .jobs import *
